+++ slang/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.5.0
 
+- feat: add `dart run slang configure` to configure `CFBundleLocalizations` in `Info.plist` based on existing translation files
 - feat: add global `generate_enum` config (#283)
 - fix: add `unused_element_parameter` to avoid warnings in Dart 3.7 (#282)